refactor: update instances of expenses, incomes, and investments to be expense, income, and investment; update dashboard widgets
