+++ src/pages/dashboard/DashboardPage.scss
@@ -21,6 +21,7 @@
     font-family: $secondary-font;
 
     .header {
+      text-decoration: none;
       font-size: 1rem;
       font-weight: 500;
       margin: 0;
